Foi criado as classes de testes para os repositórios e criado um método auxiliar para criar valores na lista
