+++ Cod3rsGrowth.Dominio/Entidades/EnumRamoDaEmpresa.cs
