+++ Cod3rsGrowth.Testes/Injecao/ModuloDeInjecao.cs
@@ -1,7 +1,6 @@
 ﻿using Cod3rsGrowth.Dominio.Interfaces;
 using Cod3rsGrowth.Dominio.Servicos;
 using Cod3rsGrowth.Infra.Interfaces;
-using Cod3rsGrowth.Testes.Mocks;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Cod3rsGrowth.Testes
