+++ Cod3rsGrowth.Testes/Mocks/EmpresaRepositorioMock.cs
@@ -1,6 +1,5 @@
 ﻿using Cod3rsGrowth.Dominio.Entidades;
 using Cod3rsGrowth.Infra.Interfaces;
-using Cod3rsGrowth.Infra.Singleton;
 
 namespace Cod3rsGrowth.Testes
 {
@@ -30,35 +29,6 @@ namespace Cod3rsGrowth.Testes
         public List<Empresa> ObterTodos()
         {
             throw new NotImplementedException();
-        }
-     
-        public void CriarLista()
-        {
-            var listaEmpresa = new List<Empresa>
-            {
-                new Empresa
-                {
-                   Id = 1,
-                   RazaoSocial = "Invent Software",
-                   CNPJ = "123456789",
-                   Ramo = EnumRamoDaEmpresa.Servico
-                },
-                new Empresa
-                {
-                   Id = 2,
-                   RazaoSocial = "Heinz",
-                   CNPJ = "987654321",
-                   Ramo = EnumRamoDaEmpresa.Industria
-                },
-                new Empresa
-                {
-                   Id = 3,
-                   RazaoSocial = "Lojas Americanas",
-                   CNPJ = "543216789",
-                   Ramo = EnumRamoDaEmpresa.Comercio
-                },
-            };
-            EmpresaSingleton.Instancia.AddRange(listaEmpresa);
         }
     }
 }
